+++ src/backend/access/common/printtup.c
@@ -431,7 +431,7 @@ printatt(unsigned attributeId,
 		   value != NULL ? " = \"" : "",
 		   value != NULL ? value : "",
 		   value != NULL ? "\"" : "",
-		   (unsigned int) (attributeP->atttypid),
+		   attributeP->atttypid,
 		   attributeP->attlen,
 		   attributeP->atttypmod,
 		   attributeP->attbyval ? 't' : 'f');
